+++ node-repository/src/test/java/com/yahoo/vespa/hosted/provision/restapi/v2/responses/maintenance.json
@@ -31,9 +31,6 @@
       "name":"ProvisionedExpirer"
     },
     {
-      "name":"RetiredEarlyExpirer"
-    },
-    {
       "name":"MetricsReporter"
     },
     {
